Samle statustall-typar i ei eiga fil
