+++ src/enhetsportefolje/enhet-side.tsx
@@ -38,7 +38,8 @@ import {lukkFeilTiltakModal} from '../ducks/lagret-filter-ui-state';
 import {Alert} from '@navikt/ds-react';
 import {Informasjonsmeldinger} from '../components/informasjonsmeldinger/informasjonsmeldinger';
 import {useStatustallEnhetSelector} from '../hooks/redux/use-statustall';
-import {StatustallEnhet, StatustallEnhetState} from '../ducks/statustall/statustall-enhet';
+import {StatustallEnhetState} from '../ducks/statustall/statustall-enhet';
+import {StatustallEnhet} from '../ducks/statustall/statustall-typer';
 
 export function antallFilter(filtervalg) {
     function mapAktivitetFilter(value) {
